requirements.txt
flask==3.1.0
flask-mysqldb==2.0.0
google-cloud-secret-manager==2.23.2
numpy==2.2.4
pandas==2.2.3
werkzeug==3.1.3